+++ private_dot_dsh/exact_profiles/exact_web/exact_plugins/exact_dsh-web-ui-notify/exact_src/private_types.d.ts
@@ -1,0 +1,6 @@
+declare module 'web-push'
+declare module '@deepseek-ai/dsh-client-runtime'
+declare module '@deepseek-ai/dsh-client-runtime/client'
+declare module '@deepseek-ai/dsh-client-ui-slots'
+declare module '@deepseek-ai/dsh-client-locale'
+declare module '@deepseek-ai/dsh-client-ui-settings'